+++ android-networking/src/main/java/com/androidnetworking/common/ANRequest.java
@@ -40,6 +40,8 @@ import com.androidnetworking.interfaces.StringRequestListener;
 import com.androidnetworking.interfaces.UploadProgressListener;
 import com.androidnetworking.internal.ANRequestQueue;
 import com.androidnetworking.internal.SynchronousCall;
+import com.androidnetworking.model.MultipartFileBody;
+import com.androidnetworking.model.MultipartStringBody;
 import com.androidnetworking.utils.ParseUtil;
 import com.androidnetworking.utils.Utils;
 import com.google.gson.internal.$Gson$Types;
@@ -50,9 +52,11 @@ import org.json.JSONObject;
 
 import java.io.File;
 import java.lang.reflect.Type;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.Executor;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
@@ -84,13 +88,13 @@ public class ANRequest<T extends ANRequest> {
     private int sequenceNumber;
     private Object mTag;
     private ResponseType mResponseType;
-    private HashMap<String, String> mHeadersMap = new HashMap<>();
+    private HashMap<String, List<String>> mHeadersMap = new HashMap<>();
     private HashMap<String, String> mBodyParameterMap = new HashMap<>();
     private HashMap<String, String> mUrlEncodedFormBodyParameterMap = new HashMap<>();
-    private HashMap<String, String> mMultiPartParameterMap = new HashMap<>();
-    private HashMap<String, String> mQueryParameterMap = new HashMap<>();
+    private HashMap<String, MultipartStringBody> mMultiPartParameterMap = new HashMap<>();
+    private HashMap<String, List<String>> mQueryParameterMap = new HashMap<>();
     private HashMap<String, String> mPathParameterMap = new HashMap<>();
-    private HashMap<String, File> mMultiPartFileMap = new HashMap<>();
+    private HashMap<String, List<MultipartFileBody>> mMultiPartFileMap = new HashMap<>();
     private String mDirPath;
     private String mFileName;
     private String mApplicationJsonString = null;
@@ -109,6 +113,7 @@ public class ANRequest<T extends ANRequest> {
     private int mProgress;
     private boolean isCancelled;
     private boolean isDelivered;
+    private boolean isRunning;
     private int mPercentageThresholdForCancelling = 0;
     private JSONArrayRequestListener mJSONArrayRequestListener;
     private JSONObjectRequestListener mJSONObjectRequestListener;
@@ -405,8 +410,17 @@ public class ANRequest<T extends ANRequest> {
             tempUrl = tempUrl.replace("{" + entry.getKey() + "}", String.valueOf(entry.getValue()));
         }
         HttpUrl.Builder urlBuilder = HttpUrl.parse(tempUrl).newBuilder();
-        for (HashMap.Entry<String, String> entry : mQueryParameterMap.entrySet()) {
-            urlBuilder.addQueryParameter(entry.getKey(), entry.getValue());
+        if (mQueryParameterMap != null) {
+            Set<Map.Entry<String, List<String>>> entries = mQueryParameterMap.entrySet();
+            for (Map.Entry<String, List<String>> entry : entries) {
+                String name = entry.getKey();
+                List<String> list = entry.getValue();
+                if (list != null) {
+                    for (String value : list) {
+                        urlBuilder.addQueryParameter(name, value);
+                    }
+                }
+            }
         }
         return urlBuilder.build().toString();
     }
@@ -537,6 +551,7 @@ public class ANRequest<T extends ANRequest> {
             if (forceCancel || mPercentageThresholdForCancelling == 0
                     || mProgress < mPercentageThresholdForCancelling) {
                 isCancelled = true;
+                isRunning = false;
                 if (call != null) {
                     call.cancel();
                 }
@@ -554,6 +569,14 @@ public class ANRequest<T extends ANRequest> {
 
     public boolean isCanceled() {
         return isCancelled;
+    }
+
+    public boolean isRunning() {
+        return isRunning;
+    }
+
+    public void setRunning(boolean running) {
+        isRunning = running;
     }
 
     public Call getCall() {
@@ -574,7 +597,7 @@ public class ANRequest<T extends ANRequest> {
 
     public void destroy() {
         mJSONArrayRequestListener = null;
-        mJSONArrayRequestListener = null;
+        mJSONObjectRequestListener = null;
         mStringRequestListener = null;
         mBitmapRequestListener = null;
         mParsedRequestListener = null;
@@ -828,22 +851,34 @@ public class ANRequest<T extends ANRequest> {
     }
 
     public RequestBody getMultiPartRequestBody() {
-        MultipartBody.Builder builder = new MultipartBody
-                .Builder()
+        MultipartBody.Builder builder = new MultipartBody.Builder()
                 .setType((customMediaType == null) ? MultipartBody.FORM : customMediaType);
         try {
-            for (HashMap.Entry<String, String> entry : mMultiPartParameterMap.entrySet()) {
+            for (HashMap.Entry<String, MultipartStringBody> entry : mMultiPartParameterMap.entrySet()) {
+                MultipartStringBody stringBody = entry.getValue();
+                MediaType mediaType = null;
+                if (stringBody.contentType != null) {
+                    mediaType = MediaType.parse(stringBody.contentType);
+                }
                 builder.addPart(Headers.of("Content-Disposition",
                         "form-data; name=\"" + entry.getKey() + "\""),
-                        RequestBody.create(null, entry.getValue()));
+                        RequestBody.create(mediaType, stringBody.value));
             }
-            for (HashMap.Entry<String, File> entry : mMultiPartFileMap.entrySet()) {
-                String fileName = entry.getValue().getName();
-                RequestBody fileBody = RequestBody.create(MediaType.parse(Utils.getMimeType(fileName)),
-                        entry.getValue());
-                builder.addPart(Headers.of("Content-Disposition",
-                        "form-data; name=\"" + entry.getKey() + "\"; filename=\"" + fileName + "\""),
-                        fileBody);
+            for (HashMap.Entry<String, List<MultipartFileBody>> entry : mMultiPartFileMap.entrySet()) {
+                List<MultipartFileBody> fileBodies = entry.getValue();
+                for (MultipartFileBody fileBody : fileBodies) {
+                    String fileName = fileBody.file.getName();
+                    MediaType mediaType;
+                    if (fileBody.contentType != null) {
+                        mediaType = MediaType.parse(fileBody.contentType);
+                    } else {
+                        mediaType = MediaType.parse(Utils.getMimeType(fileName));
+                    }
+                    RequestBody requestBody = RequestBody.create(mediaType, fileBody.file);
+                    builder.addPart(Headers.of("Content-Disposition",
+                            "form-data; name=\"" + entry.getKey() + "\"; filename=\"" + fileName + "\""),
+                            requestBody);
+                }
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -854,8 +889,17 @@ public class ANRequest<T extends ANRequest> {
     public Headers getHeaders() {
         Headers.Builder builder = new Headers.Builder();
         try {
-            for (HashMap.Entry<String, String> entry : mHeadersMap.entrySet()) {
-                builder.add(entry.getKey(), entry.getValue());
+            if (mHeadersMap != null) {
+                Set<Map.Entry<String, List<String>>> entries = mHeadersMap.entrySet();
+                for (Map.Entry<String, List<String>> entry : entries) {
+                    String name = entry.getKey();
+                    List<String> list = entry.getValue();
+                    if (list != null) {
+                        for (String value : list) {
+                            builder.add(name, value);
+                        }
+                    }
+                }
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -870,6 +914,13 @@ public class ANRequest<T extends ANRequest> {
         }
     }
 
+    public static class OptionsRequestBuilder extends GetRequestBuilder {
+
+        public OptionsRequestBuilder(String url) {
+            super(url, Method.OPTIONS);
+        }
+    }
+
     public static class GetRequestBuilder<T extends GetRequestBuilder> implements RequestBuilder {
         private Priority mPriority = Priority.MEDIUM;
         private int mMethod = Method.GET;
@@ -880,8 +931,8 @@ public class ANRequest<T extends ANRequest> {
         private int mMaxWidth;
         private int mMaxHeight;
         private ImageView.ScaleType mScaleType;
-        private HashMap<String, String> mHeadersMap = new HashMap<>();
-        private HashMap<String, String> mQueryParameterMap = new HashMap<>();
+        private HashMap<String, List<String>> mHeadersMap = new HashMap<>();
+        private HashMap<String, List<String>> mQueryParameterMap = new HashMap<>();
         private HashMap<String, String> mPathParameterMap = new HashMap<>();
         private CacheControl mCacheControl;
         private Executor mExecutor;
@@ -912,14 +963,23 @@ public class ANRequest<T extends ANRequest> {
 
         @Override
         public T addQueryParameter(String key, String value) {
-            mQueryParameterMap.put(key, value);
+            List<String> list = mQueryParameterMap.get(key);
+            if (list == null) {
+                list = new ArrayList<>();
+                mQueryParameterMap.put(key, list);
+            }
+            if (!list.contains(value)) {
+                list.add(value);
+            }
             return (T) this;
         }
 
         @Override
         public T addQueryParameter(Map<String, String> queryParameterMap) {
             if (queryParameterMap != null) {
-                mQueryParameterMap.putAll(queryParameterMap);
+                for (HashMap.Entry<String, String> entry : queryParameterMap.entrySet()) {
+                    addQueryParameter(entry.getKey(), entry.getValue());
+                }
             }
             return (T) this;
         }
@@ -927,7 +987,7 @@ public class ANRequest<T extends ANRequest> {
         @Override
         public T addQueryParameter(Object object) {
             if (object != null) {
-                mQueryParameterMap.putAll(ParseUtil
+                return addQueryParameter(ParseUtil
                         .getParserFactory()
                         .getStringMap(object));
             }
@@ -960,14 +1020,23 @@ public class ANRequest<T extends ANRequest> {
 
         @Override
         public T addHeaders(String key, String value) {
-            mHeadersMap.put(key, value);
+            List<String> list = mHeadersMap.get(key);
+            if (list == null) {
+                list = new ArrayList<>();
+                mHeadersMap.put(key, list);
+            }
+            if (!list.contains(value)) {
+                list.add(value);
+            }
             return (T) this;
         }
 
         @Override
         public T addHeaders(Map<String, String> headerMap) {
             if (headerMap != null) {
-                mHeadersMap.putAll(headerMap);
+                for (HashMap.Entry<String, String> entry : headerMap.entrySet()) {
+                    addHeaders(entry.getKey(), entry.getValue());
+                }
             }
             return (T) this;
         }
@@ -975,7 +1044,7 @@ public class ANRequest<T extends ANRequest> {
         @Override
         public T addHeaders(Object object) {
             if (object != null) {
-                mHeadersMap.putAll(ParseUtil
+                return addHeaders(ParseUtil
                         .getParserFactory()
                         .getStringMap(object));
             }
@@ -1081,6 +1150,13 @@ public class ANRequest<T extends ANRequest> {
         }
     }
 
+    public static class DynamicRequestBuilder extends PostRequestBuilder {
+
+        public DynamicRequestBuilder(String url, int method) {
+            super(url, method);
+        }
+    }
+
     public static class PostRequestBuilder<T extends PostRequestBuilder> implements RequestBuilder {
 
         private Priority mPriority = Priority.MEDIUM;
@@ -1091,10 +1167,10 @@ public class ANRequest<T extends ANRequest> {
         private String mStringBody = null;
         private byte[] mByte = null;
         private File mFile = null;
-        private HashMap<String, String> mHeadersMap = new HashMap<>();
+        private HashMap<String, List<String>> mHeadersMap = new HashMap<>();
         private HashMap<String, String> mBodyParameterMap = new HashMap<>();
         private HashMap<String, String> mUrlEncodedFormBodyParameterMap = new HashMap<>();
-        private HashMap<String, String> mQueryParameterMap = new HashMap<>();
+        private HashMap<String, List<String>> mQueryParameterMap = new HashMap<>();
         private HashMap<String, String> mPathParameterMap = new HashMap<>();
         private CacheControl mCacheControl;
         private Executor mExecutor;
@@ -1126,14 +1202,23 @@ public class ANRequest<T extends ANRequest> {
 
         @Override
         public T addQueryParameter(String key, String value) {
-            mQueryParameterMap.put(key, value);
+            List<String> list = mQueryParameterMap.get(key);
+            if (list == null) {
+                list = new ArrayList<>();
+                mQueryParameterMap.put(key, list);
+            }
+            if (!list.contains(value)) {
+                list.add(value);
+            }
             return (T) this;
         }
 
         @Override
         public T addQueryParameter(Map<String, String> queryParameterMap) {
             if (queryParameterMap != null) {
-                mQueryParameterMap.putAll(queryParameterMap);
+                for (HashMap.Entry<String, String> entry : queryParameterMap.entrySet()) {
+                    addQueryParameter(entry.getKey(), entry.getValue());
+                }
             }
             return (T) this;
         }
@@ -1141,7 +1226,7 @@ public class ANRequest<T extends ANRequest> {
         @Override
         public T addQueryParameter(Object object) {
             if (object != null) {
-                mQueryParameterMap.putAll(ParseUtil
+                return addQueryParameter(ParseUtil
                         .getParserFactory()
                         .getStringMap(object));
             }
@@ -1174,14 +1259,23 @@ public class ANRequest<T extends ANRequest> {
 
         @Override
         public T addHeaders(String key, String value) {
-            mHeadersMap.put(key, value);
+            List<String> list = mHeadersMap.get(key);
+            if (list == null) {
+                list = new ArrayList<>();
+                mHeadersMap.put(key, list);
+            }
+            if (!list.contains(value)) {
+                list.add(value);
+            }
             return (T) this;
         }
 
         @Override
         public T addHeaders(Map<String, String> headerMap) {
             if (headerMap != null) {
-                mHeadersMap.putAll(headerMap);
+                for (HashMap.Entry<String, String> entry : headerMap.entrySet()) {
+                    addHeaders(entry.getKey(), entry.getValue());
+                }
             }
             return (T) this;
         }
@@ -1189,7 +1283,7 @@ public class ANRequest<T extends ANRequest> {
         @Override
         public T addHeaders(Object object) {
             if (object != null) {
-                mHeadersMap.putAll(ParseUtil
+                return addHeaders(ParseUtil
                         .getParserFactory()
                         .getStringMap(object));
             }
@@ -1339,8 +1433,8 @@ public class ANRequest<T extends ANRequest> {
         private Priority mPriority = Priority.MEDIUM;
         private String mUrl;
         private Object mTag;
-        private HashMap<String, String> mHeadersMap = new HashMap<>();
-        private HashMap<String, String> mQueryParameterMap = new HashMap<>();
+        private HashMap<String, List<String>> mHeadersMap = new HashMap<>();
+        private HashMap<String, List<String>> mQueryParameterMap = new HashMap<>();
         private HashMap<String, String> mPathParameterMap = new HashMap<>();
         private String mDirPath;
         private String mFileName;
@@ -1370,14 +1464,23 @@ public class ANRequest<T extends ANRequest> {
 
         @Override
         public T addHeaders(String key, String value) {
-            mHeadersMap.put(key, value);
+            List<String> list = mHeadersMap.get(key);
+            if (list == null) {
+                list = new ArrayList<>();
+                mHeadersMap.put(key, list);
+            }
+            if (!list.contains(value)) {
+                list.add(value);
+            }
             return (T) this;
         }
 
         @Override
         public T addHeaders(Map<String, String> headerMap) {
             if (headerMap != null) {
-                mHeadersMap.putAll(headerMap);
+                for (HashMap.Entry<String, String> entry : headerMap.entrySet()) {
+                    addHeaders(entry.getKey(), entry.getValue());
+                }
             }
             return (T) this;
         }
@@ -1385,7 +1488,7 @@ public class ANRequest<T extends ANRequest> {
         @Override
         public T addHeaders(Object object) {
             if (object != null) {
-                mHeadersMap.putAll(ParseUtil
+                return addHeaders(ParseUtil
                         .getParserFactory()
                         .getStringMap(object));
             }
@@ -1394,14 +1497,23 @@ public class ANRequest<T extends ANRequest> {
 
         @Override
         public T addQueryParameter(String key, String value) {
-            mQueryParameterMap.put(key, value);
+            List<String> list = mQueryParameterMap.get(key);
+            if (list == null) {
+                list = new ArrayList<>();
+                mQueryParameterMap.put(key, list);
+            }
+            if (!list.contains(value)) {
+                list.add(value);
+            }
             return (T) this;
         }
 
         @Override
         public T addQueryParameter(Map<String, String> queryParameterMap) {
             if (queryParameterMap != null) {
-                mQueryParameterMap.putAll(queryParameterMap);
+                for (HashMap.Entry<String, String> entry : queryParameterMap.entrySet()) {
+                    addQueryParameter(entry.getKey(), entry.getValue());
+                }
             }
             return (T) this;
         }
@@ -1409,7 +1521,7 @@ public class ANRequest<T extends ANRequest> {
         @Override
         public T addQueryParameter(Object object) {
             if (object != null) {
-                mQueryParameterMap.putAll(ParseUtil
+                return addQueryParameter(ParseUtil
                         .getParserFactory()
                         .getStringMap(object));
             }
@@ -1503,11 +1615,11 @@ public class ANRequest<T extends ANRequest> {
         private Priority mPriority = Priority.MEDIUM;
         private String mUrl;
         private Object mTag;
-        private HashMap<String, String> mHeadersMap = new HashMap<>();
-        private HashMap<String, String> mMultiPartParameterMap = new HashMap<>();
-        private HashMap<String, String> mQueryParameterMap = new HashMap<>();
+        private HashMap<String, List<String>> mHeadersMap = new HashMap<>();
+        private HashMap<String, List<String>> mQueryParameterMap = new HashMap<>();
         private HashMap<String, String> mPathParameterMap = new HashMap<>();
-        private HashMap<String, File> mMultiPartFileMap = new HashMap<>();
+        private HashMap<String, MultipartStringBody> mMultiPartParameterMap = new HashMap<>();
+        private HashMap<String, List<MultipartFileBody>> mMultiPartFileMap = new HashMap<>();
         private CacheControl mCacheControl;
         private int mPercentageThresholdForCancelling = 0;
         private Executor mExecutor;
@@ -1533,14 +1645,23 @@ public class ANRequest<T extends ANRequest> {
 
         @Override
         public T addQueryParameter(String key, String value) {
-            mQueryParameterMap.put(key, value);
+            List<String> list = mQueryParameterMap.get(key);
+            if (list == null) {
+                list = new ArrayList<>();
+                mQueryParameterMap.put(key, list);
+            }
+            if (!list.contains(value)) {
+                list.add(value);
+            }
             return (T) this;
         }
 
         @Override
         public T addQueryParameter(Map<String, String> queryParameterMap) {
             if (queryParameterMap != null) {
-                mQueryParameterMap.putAll(queryParameterMap);
+                for (HashMap.Entry<String, String> entry : queryParameterMap.entrySet()) {
+                    addQueryParameter(entry.getKey(), entry.getValue());
+                }
             }
             return (T) this;
         }
@@ -1548,7 +1669,7 @@ public class ANRequest<T extends ANRequest> {
         @Override
         public T addQueryParameter(Object object) {
             if (object != null) {
-                mQueryParameterMap.putAll(ParseUtil
+                return addQueryParameter(ParseUtil
                         .getParserFactory()
                         .getStringMap(object));
             }
@@ -1581,14 +1702,23 @@ public class ANRequest<T extends ANRequest> {
 
         @Override
         public T addHeaders(String key, String value) {
-            mHeadersMap.put(key, value);
+            List<String> list = mHeadersMap.get(key);
+            if (list == null) {
+                list = new ArrayList<>();
+                mHeadersMap.put(key, list);
+            }
+            if (!list.contains(value)) {
+                list.add(value);
+            }
             return (T) this;
         }
 
         @Override
         public T addHeaders(Map<String, String> headerMap) {
             if (headerMap != null) {
-                mHeadersMap.putAll(headerMap);
+                for (HashMap.Entry<String, String> entry : headerMap.entrySet()) {
+                    addHeaders(entry.getKey(), entry.getValue());
+                }
             }
             return (T) this;
         }
@@ -1596,7 +1726,7 @@ public class ANRequest<T extends ANRequest> {
         @Override
         public T addHeaders(Object object) {
             if (object != null) {
-                mHeadersMap.putAll(ParseUtil
+                return addHeaders(ParseUtil
                         .getParserFactory()
                         .getStringMap(object));
             }
@@ -1652,34 +1782,100 @@ public class ANRequest<T extends ANRequest> {
         }
 
         public T addMultipartParameter(String key, String value) {
-            mMultiPartParameterMap.put(key, value);
+            return addMultipartParameter(key, value, null);
+        }
+
+        public T addMultipartParameter(String key, String value, String contentType) {
+            MultipartStringBody stringBody = new MultipartStringBody(value, contentType);
+            mMultiPartParameterMap.put(key, stringBody);
             return (T) this;
         }
 
         public T addMultipartParameter(Map<String, String> multiPartParameterMap) {
+            return addMultipartParameter(multiPartParameterMap, null);
+        }
+
+        public T addMultipartParameter(Map<String, String> multiPartParameterMap, String contentType) {
             if (multiPartParameterMap != null) {
-                mMultiPartParameterMap.putAll(multiPartParameterMap);
+                Map<String, MultipartStringBody> parameterMap = new HashMap<>();
+                for (HashMap.Entry<String, String> entry : multiPartParameterMap.entrySet()) {
+                    MultipartStringBody stringBody = new MultipartStringBody(entry.getValue(), contentType);
+                    parameterMap.put(entry.getKey(), stringBody);
+                }
+                mMultiPartParameterMap.putAll(parameterMap);
             }
             return (T) this;
         }
 
         public T addMultipartParameter(Object object) {
+            return addMultipartParameter(object, null);
+        }
+
+        public T addMultipartParameter(Object object, String contentType) {
             if (object != null) {
-                mMultiPartParameterMap.putAll(ParseUtil
+                Map<String, String> parameterMap = ParseUtil
                         .getParserFactory()
-                        .getStringMap(object));
+                        .getStringMap(object);
+                addMultipartParameter(parameterMap, contentType);
             }
             return (T) this;
         }
 
         public T addMultipartFile(String key, File file) {
-            mMultiPartFileMap.put(key, file);
+            return addMultipartFile(key, file, null);
+        }
+
+        public T addMultipartFile(String key, File file, String contentType) {
+            MultipartFileBody fileBody = new MultipartFileBody(file, contentType);
+            addMultipartFileWithKey(key, fileBody);
             return (T) this;
         }
 
         public T addMultipartFile(Map<String, File> multiPartFileMap) {
+            return addMultipartFile(multiPartFileMap, null);
+        }
+
+        public T addMultipartFile(Map<String, File> multiPartFileMap, String contentType) {
             if (multiPartFileMap != null) {
-                mMultiPartFileMap.putAll(multiPartFileMap);
+                for (HashMap.Entry<String, File> entry : multiPartFileMap.entrySet()) {
+                    MultipartFileBody fileBody = new MultipartFileBody(entry.getValue(), contentType);
+                    addMultipartFileWithKey(entry.getKey(), fileBody);
+                }
+            }
+            return (T) this;
+        }
+
+        public T addMultipartFileList(String key, List<File> files) {
+            return addMultipartFileList(key, files, null);
+        }
+
+        public T addMultipartFileList(String key, List<File> files, String contentType) {
+            if (files != null) {
+                for (File file : files) {
+                    MultipartFileBody fileBody = new MultipartFileBody(file, contentType);
+                    addMultipartFileWithKey(key, fileBody);
+                }
+            }
+            return (T) this;
+        }
+
+        public T addMultipartFileList(Map<String, List<File>> multiPartFileMap) {
+            return addMultipartFileList(multiPartFileMap, null);
+        }
+
+        public T addMultipartFileList(Map<String, List<File>> multiPartFileMap, String contentType) {
+            if (multiPartFileMap != null) {
+                Map<String, List<MultipartFileBody>> parameterMap = new HashMap<>();
+                for (HashMap.Entry<String, List<File>> entry : multiPartFileMap.entrySet()) {
+                    List<File> files = entry.getValue();
+                    List<MultipartFileBody> fileBodies = new ArrayList<>();
+                    for (File file : files) {
+                        MultipartFileBody fileBody = new MultipartFileBody(file, contentType);
+                        fileBodies.add(fileBody);
+                    }
+                    parameterMap.put(entry.getKey(), fileBodies);
+                }
+                mMultiPartFileMap.putAll(parameterMap);
             }
             return (T) this;
         }
@@ -1692,6 +1888,15 @@ public class ANRequest<T extends ANRequest> {
         public T setContentType(String contentType) {
             mCustomContentType = contentType;
             return (T) this;
+        }
+
+        private void addMultipartFileWithKey(String key, MultipartFileBody fileBody) {
+            List<MultipartFileBody> fileBodies = mMultiPartFileMap.get(key);
+            if (fileBodies == null) {
+                fileBodies = new ArrayList<>();
+            }
+            fileBodies.add(fileBody);
+            mMultiPartFileMap.put(key, fileBodies);
         }
 
         public ANRequest build() {